revisions after first session
I introduced better (I hope!) explanations about how a function is constructed, what it does, and how it is used.
I also got rid of the "my_mean" function example, because I felt it confused the students. Instead, I introduced a function that computes the percentage of a number of a given total.